fix up data layer tests
