+++ packages/data-layer/package.json
@@ -37,8 +37,8 @@
     "clean": "npx rimraf dist coverage",
     "clean:deps": "npx rimraf node_modules",
     "compliance": "yarn run lint && yarn run test && yarn run coverage",
-    "test": "true",
-    "coverage": "true",
+    "test": "c8 -x src/bootstrap -x \"src/**/*.test.*\" mocha -r ts-node/register src/*.test.ts src/**/*.test.ts",
+    "coverage": "c8 check-coverage --statements 90 --branches 90 --functions 90 --lines 90",
     "lint": "eslint \"**/*.ts\"",
     "lint:fix": "eslint --fix \"**/*.ts\"",
     "build": "tsc"
